requirements.txt
Django==1.6.2

dj-database-url==0.2.2
dj-static==0.0.5
django-cache-url==0.7.0
django-cors-headers==0.12
djangorestframework==2.3.12
gunicorn==18.0
pygeocoder==1.2.4
psycopg2==2.5.2
raven==4.0.4
South==0.8.4
sqlparse==0.1.11
wsgiref==0.1.2
